Reduxed select and autofill routine in newWorkout page
