+++ client/src/components/new-workout/NewWorkoutHeader.jsx
@@ -1,12 +1,12 @@
 import RoutineDropdown from "../routine-dropdown/RoutineDropdown";
 import DateDropdown from "../new-workout/DateDropdown";
 
-const NewWorkoutHeader = ({ chooseRoutine, workoutData }) => {
+const NewWorkoutHeader = ({ workoutData }) => {
   return (
     <>
       <h2 className="text-left text-2xl ">New Workout</h2>
       <div className="workout-header pb-5">
-        <RoutineDropdown chooseRoutine={chooseRoutine} />
+        <RoutineDropdown />
         <DateDropdown workoutData={workoutData} />
       </div>
     </>
