+++ client/src/containers/NewWorkout.jsx
@@ -1,13 +1,11 @@
-import { useSelector, useDispatch } from "react-redux";
+import { useSelector } from "react-redux";
 import ExerciseList from "../components/new-workout/ExerciseList";
 import NewWorkoutHeader from "../components/new-workout/NewWorkoutHeader";
 import NewWorkoutFooter from "../components/new-workout/NewWorkoutFooter";
-import { setWorkoutData } from "../utils/slice/newWorkoutSlice";
 
-const NewWorkout = ({ updatePage }) => {
+const NewWorkout = () => {
   const { workoutData } = useSelector((state) => state.newWorkout);
-  const { routines } = useSelector((state) => state.routine);
-  const dispatch = useDispatch();
+
   // const validator = () => {
   //   const body = { ...workoutData, exercises: [] };
   //   if (workoutData.date) {
@@ -42,31 +40,9 @@ const NewWorkout = ({ updatePage }) => {
   //   return true;
   // };
 
-  /* CHOOSE ROUTINE AND RENDER */
-  const chooseRoutine = (e) => {
-    /* REFACTOR TO USE ROUTINE ID */
-    const selectedRoutine =
-      routines && routines.find((routine) => routine.name === e.target.value);
-    const data = {
-      routine: selectedRoutine.name,
-      exercises: selectedRoutine.data.map(({ exercise, sets }) => {
-        const data = Array.from({ length: sets }, () => ({
-          reps: null,
-          weight: null,
-          notes: null,
-        }));
-        return { name: exercise, data };
-      }),
-    };
-    dispatch(setWorkoutData(data));
-  };
-
   return (
     <div className="pl-5 pr-5">
-      <NewWorkoutHeader
-        chooseRoutine={chooseRoutine}
-        workoutData={workoutData}
-      />
+      <NewWorkoutHeader workoutData={workoutData} />
       <ExerciseList exercises={workoutData.exercises} />
       <NewWorkoutFooter workoutData={workoutData} />
     </div>
